refactor(skill-svc): Layer 2 — extract config/mapping layer into _config.py
Zero-behavior-change extraction of the config/mapping resolution layer from
skill_progression_service.py into app/services/skill_progression/_config.py.

Moved to _config.py (minimal DB access, no EMA logic):
  - get_all_skill_keys()           — iterates SKILL_CATEGORIES; no DB
  - get_baseline_skills()          — 1 UserLicense query; handles two JSON formats
  - _extract_tournament_skills()   — reward_config priority path + TournamentSkillMapping fallback

skill_progression_service.py (1,129 → 1,017 lines) re-exports all three symbols
via the existing .skill_progression package import block. All mock.patch paths
at "app.services.skill_progression_service.*" continue to work: Layer 4/5
callers still live in the shim and resolve these names via LOAD_GLOBAL against
the shim's __dict__.

Note: get_all_skill_keys() in skills_config.py is a separate implementation
that produces identical results; the two are intentionally kept distinct.
